requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
pydantic>=2.6
numpy>=1.24
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
pdfplumber>=0.11
python-docx>=1.1.0
anthropic>=0.76.0
python-multipart>=0.0.9
streamlit>=1.31.0
requests>=2.31.0
